Laser pinwheel touchup

+++ Content/Items/Weapons/Ranger/MagmaMortar.cs
@@ -236,7 +236,7 @@ namespace ExoriumMod.Content.Items.Weapons.Ranger
             if (Projectile.ai[0] == 1)
                 Projectile.rotation = MathHelper.Pi;
 
-            Lighting.AddLight(Projectile.Center, .75f * (255 - Projectile.alpha / 255), .45f * (255 - Projectile.alpha / 255), 0);
+            Lighting.AddLight(Projectile.Center, .75f * (float)((255 - Projectile.alpha) / 255), .45f * (float)((255 - Projectile.alpha) / 255), 0);
         }
 
         public override void OnHitNPC(NPC target, NPC.HitInfo hit, int damageDone)
